[A2Y-177] space 모듈로 변경
