+++ .idea/gradle.xml
@@ -17,7 +17,7 @@
             <option value="$PROJECT_DIR$/feature" />
             <option value="$PROJECT_DIR$/feature/common" />
             <option value="$PROJECT_DIR$/feature/home" />
-            <option value="$PROJECT_DIR$/feature/locker" />
+            <option value="$PROJECT_DIR$/feature/space" />
             <option value="$PROJECT_DIR$/testbase" />
           </set>
         </option>
