+++ app/build.gradle
@@ -38,6 +38,6 @@ dependencies {
     implementation "androidx.core:core-splashscreen:$androidCoreSplashScreenVersion"
     implementation project(':feature:home')
     implementation project(':data')
-    implementation project(':feature:locker')
+    implementation project(':feature:space')
 
 }
