+++ feature/home/build.gradle
@@ -15,9 +15,8 @@ android {
 }
 
 dependencies {
-    implementation project(':feature:locker')
-
-implementation 'androidx.appcompat:appcompat:1.5.1'
+    implementation project(':feature:space')
+    implementation 'androidx.appcompat:appcompat:1.5.1'
     implementation 'com.google.android.material:material:1.4.+'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
 }
