--- feature/locker/src/main/AndroidManifest.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.yapp.itemfinder.locker">
-    <application>
-
-    </application>
-
-</manifest>
