--- feature/locker/src/main/res/values/strings.xml
@@ -1,2 +1,0 @@
-<resources>
-</resources>
