--- feature/locker/src/test/java/com/yapp/itemfinder/locker/ExampleUnitTest.kt
@@ -1,17 +1,0 @@
-package com.yapp.itemfinder.locker
-
-import org.junit.Test
-
-import org.junit.Assert.*
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * See [testing documentation](http://d.android.com/tools/testing).
- */
-class ExampleUnitTest {
-    @Test
-    fun addition_isCorrect() {
-        assertEquals(4, 2 + 2)
-    }
-}
