+++ feature/space/src/main/AndroidManifest.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.yapp.itemfinder.space">
+    <application>
+
+    </application>
+
+</manifest>
