+++ settings.gradle
@@ -27,4 +27,4 @@ include ':data'
 include ':domain'
 include ':feature:common'
 include ':feature:home'
-include ':feature:locker'
+include ':feature:space'
